+++ src/main.rs
@@ -63,10 +63,12 @@ fn is_valid_reward_multiplier(reward_mult: String) -> Result<(), String> {
 fn is_valid_bundle_size(bundle_size: String) -> Result<(), String> {
     match bundle_size.parse::<u64>() {
         Ok(n) => {
-            if n <= 200000000 {
+            if n > 0 && n <= 200 {
                 Ok(())
             } else {
-                Err(format!("Bundle data size must not be greater than 200MB."))
+                Err(format!(
+                    "Bundle size must be at least 1 MB and not greater than 200 MB."
+                ))
             }
         }
         Err(_) => Err(format!("Not a valid bundle size.")),
@@ -201,23 +203,14 @@ fn buffer_arg<'a, 'b>(default: &'a str) -> Arg<'a, 'b> {
         .help("Sets the maximum number of concurrent network requests.")
 }
 
-fn num_chunks_arg<'a, 'b>() -> Arg<'a, 'b> {
-    Arg::with_name("num_chunks")
-        .long("num-chunks")
-        .value_name("NUM_CHUNKS")
-        .takes_value(true)
-        .validator(is_parsable::<usize>)
-        .help("Sets the number of folders to chunk files into.")
-}
-
 fn bundle_size_arg<'a, 'b>() -> Arg<'a, 'b> {
     Arg::with_name("bundle_size")
         .long("bundle-size")
         .value_name("BUNDLE_SIZE")
         .takes_value(true)
         .validator(is_valid_bundle_size)
-        .default_value("10000000")
-        .help("Sets the maximum file data bytes to include in a bundle.")
+        .default_value("10")
+        .help("Sets the bundle size in megabytes.")
 }
 
 fn glob_arg<'a, 'b>(required: bool) -> Arg<'a, 'b> {
@@ -387,14 +380,6 @@ fn get_app() -> App<'static, 'static> {
                 .help("Return information in specified output format."),
         )
         .subcommand(
-            SubCommand::with_name("chunk-files")
-                .about(
-                    "Chunks files into number of folders.",
-                )
-                .arg(glob_arg(true))
-                .arg(num_chunks_arg())
-        )
-        .subcommand(
             SubCommand::with_name("estimate")
                 .about(
                     "Prints the estimated cost of uploading file(s) \
@@ -529,11 +514,6 @@ async fn main() -> CommandResult {
     let (sub_command, arg_matches) = app_matches.subcommand();
 
     match (sub_command, arg_matches) {
-        ("chunk-files", Some(sub_arg_matches)) => {
-            let glob_str = &sub_arg_matches.value_of("glob").unwrap().expand_tilde();
-            let num_chunks = value_t!(sub_arg_matches.value_of("num_chunks"), usize).unwrap();
-            command_chunk_files(glob_str, num_chunks).await
-        }
         ("balance", Some(sub_arg_matches)) => {
             let arweave = if let Some(ar_keypair_path) = sub_arg_matches.value_of("ar_keypair_path")
             {
@@ -552,7 +532,8 @@ async fn main() -> CommandResult {
             let glob_str = &sub_arg_matches.value_of("glob").unwrap().expand_tilde();
             let reward_mult = value_t!(sub_arg_matches.value_of("reward_multiplier"), f32).unwrap();
             let with_sol = sub_arg_matches.is_present("with_sol");
-            let bundle_size = value_t!(sub_arg_matches.value_of("bundle_size"), u64).unwrap();
+            let bundle_size =
+                value_t!(sub_arg_matches.value_of("bundle_size"), u64).unwrap() * 1_000_000;
             let no_bundle = sub_arg_matches.is_present("no_bundle");
             command_get_cost(
                 &Arweave::default(),
@@ -648,7 +629,8 @@ async fn main() -> CommandResult {
                 .value_of("log_dir")
                 .map(|s| s.expand_tilde());
             let reward_mult = value_t!(sub_arg_matches.value_of("reward_multiplier"), f32).unwrap();
-            let bundle_size = value_t!(sub_arg_matches.value_of("bundle_size"), u64).unwrap();
+            let bundle_size =
+                value_t!(sub_arg_matches.value_of("bundle_size"), u64).unwrap() * 1_000_000;
             let with_sol = sub_arg_matches.is_present("with_sol");
             let no_bundle = sub_arg_matches.is_present("no_bundle");
             let buffer = value_t!(sub_arg_matches.value_of("buffer"), usize).unwrap();