+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes starting with v0.1.34 to this project will be documented in 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# v0.1.40 (2021-12-10)
+- **changed:** `--bundle-size` units from bytes to megabytes.
+- **added:** handle errors from getting Arweave and oracle prices.
+
 # v0.1.38 (2021-12-09)
 - **changed:** reduced `--with-sol` cost to 10,000 lamports per transaction.